feat: Update Form component to submit form data to Formspree
The Form component was updated to submit form data to Formspree using the `fetch` API. The endpoint URL was changed to `https://formspree.io/f/mqazarvb`. If the form submission is successful, the submit success state is set to true and the form is reset. If there is an error during form submission, an error message is logged to the console.